Order: add regex to improve origin match
