+++ Cargo.toml
@@ -17,6 +17,7 @@ elsa = "1.9.0"
 lazy_static = "1.4.0"
 log = "0.4.20"
 log4rs = { version = "1.2.0", features = ["gzip", "background_rotation"] }
+regex = "1.10.2"
 rust_decimal = { version = "1.32.0", features = ["db-diesel2-mysql"] }
 serde = { version = "1.0.183", features = ["derive"] }
 url = "2.4.1"
