+++ tests/fixtures/order_for_unit_test.csv
@@ -1,3 +1,3 @@
 c_order_id;c_bpartner_id;client_name;date;order_ref;po_ref;origin;completion;order_status
-1;1;Client 1;2023-08-01;Ref1;PoRef1;Web;30;CO
+1;1;Client 1;2023-08-01;Ref1;PoRef1;Web Order;30;CO
 2;2;Client 2;2023-08-02;Ref2;PoRef2;EDI;20;IN
