+++ tests/fixtures/orders.csv
@@ -1,6 +1,6 @@
 c_bpartner_id;client_name;date;order_ref;c_order_id;po_ref;origin;completion;order_status
-1009681;Client 1;07/06/2023;OV426946;1113194;P23HA01525;Web;0.9%;IP
+1009681;Client 1;07/06/2023;OV426946;1113194;P23HA01525;Web 123;0.9%;IP
 1012489;;12/04/2023;OV426905;1112737;230000535;;100%;CO
-1010591;Client 3;17/03/2023;WEB82760;1118684;;Web;50;CO
+1010591;Client 3;17/03/2023;WEB82760;1118684;;Commande Web;50;CO
 1012519;Client 4;17/03/2023;WEB82655;1118561;CF653;Web;101%;CO
 1010265;Client 5;17/03/2023;OV427619;1118643;WEB73714;;100%;
